requirements.txt
python-telegram-bot==21.5
APScheduler==3.10.4
python-dateutil==2.9.0.post0
pytz==2024.1
python-dotenv==1.0.1
